Fixed bug with global indentation level
Now that I've fixed the bug, I can remove the concept of an indentation
level entirely.

Upon further reflection, it's not on us to dictate how much the user
can/should open & close their brackets. If they want to (or have to, as
a result of their particular config file format) mismatch their braces,
it isn't on us to restrict them.

Though I'm glad I fixed the reason for the problem. It was the right
thing to do.